+++ src/utils/application/Transaction.java
@@ -2,5 +2,5 @@ package utils.application;
 
 import java.io.Serializable;
 
-public record Transaction(Integer id, Double amount, Integer sender, Integer receiver) implements Serializable {
+public record Transaction(Long id, Double amount, Integer sender, Integer receiver) implements Serializable {
 }
